Added a patch to apply after copying Mali files
Note that the Mali tarball is terribly packaged. I really wonder if
the person who packaged this tarball had any knowledge of a Unix.
All the files and folders in the tarball have Read-Write-Execute flags.

So before copying the file, some operations need to be done once you've
unpacked the tarball. Do not unpack it directly in the kernel tree...
In the extracted tarball folder :
$ find -type 'd' -exec chmod 0755 {} ';'
$ find -type 'f' -exec chmod 0644 {} ';'
In the extracted tarball folders, you'll also have to remove the
license.txt file, the patches directory. I'll see about removing
sconscripts too...

Once these operations done, you can copy the content of the extracted
'kernel' sub-folder in the linux kernel tree.

Note that you can apply this patch before or after. Order is not
important.

Signed-off-by: Myy <myy@miouyouyou.fr>