Trying to lie to Mali
I'm beginning to wonder if this version mismatch thing is THAT
important. In order to test this, I've build a new kernel that
includes the Mali r14p0 kernel modules but report them as Mali
r12p0.

I'm going to test this with the new r12p0 User-space drivers
provided by ARM, for Firefly boards (same chipset).

I'll see how it goes.

Signed-off-by: Myy <myy@miouyouyou.fr>